feat: change 404 page logic
