+++ my-app/src/components/movieContainer/MovieContainer.jsx
@@ -2,8 +2,9 @@ import React, { useState, useEffect } from 'react';
 import { connect, useDispatch, useSelector } from 'react-redux';
 import MovieCard from '../movieCard/MovieCard';
 import './MovieContainer.scss';
-import { getMoviesAPI, setCurrentPage } from '../../redux/actions';
-import { createPages } from '../../redux/createPages';
+import { createPages } from '../../utils/createPages';
+import { getMoviesAPI } from '../../redux/api';
+import { setCurrentPage } from '../../redux/actions';
 
 const MovieContainer = () => {
   const [showOptions, setShowOptions] = useState({});
@@ -12,13 +13,14 @@ const MovieContainer = () => {
   const totalCount = useSelector(state => state.totalCount); //получаем состояние из store
   const currentPage = useSelector(state => state.currentPage); //получаем состояние из store
   const total = useSelector(state => state.total); //получаем состояние из store
+  const isFetchedError = useSelector(state => state.isFetchedError);
   const pages = [];
 
   createPages(pages, total, currentPage);
 
   const handleOpenClick = (event, id) => {
     const item = movies.filter((movie) => movie.kinopoiskId === id)[0]; // данные одного выбранного фильма
-    setShowOptions((item.kinopoiskId === id) ? { [item.kinopoiskId]: true } : { [item.kinopoiskId]: false });
+    setShowOptions((item.kinopoiskId === id) ? { [item.kinopoiskId]: true } : null);
     event.stopPropagation();
   }
 
@@ -31,7 +33,7 @@ const MovieContainer = () => {
   useEffect(() => {
     dispatch(getMoviesAPI(currentPage));
     window.scroll(0, 0);
-  }, [currentPage, dispatch]);
+  }, [currentPage]);
 
   useEffect(() => {
     document.addEventListener('click', handleCloseClick);
@@ -40,12 +42,15 @@ const MovieContainer = () => {
     }
   }, []);
 
-  if (totalCount === 0) {
-    return (
-      <div className="movie__container">
-        <div className="loading">No movies found</div>
+  if (isFetchedError) {
+    return (<div className="error">
+      <div className="error__wrapper">
+        <div className="error_content">
+          <h2 className="error_text">Page not Found</h2>
+          <p className="error_num">404</p>
+        </div>
       </div>
-    )
+    </div>)
   }
 
   if (totalCount === undefined) {
@@ -56,12 +61,21 @@ const MovieContainer = () => {
     )
   }
 
+  if (totalCount === 0) {
+    return (
+      <div className="movie__container">
+        <div className="loading">No movies found</div>
+      </div>
+    )
+  }
+
   return (
     <div className="movie__container">
       <div className="result">
         <span className="result__count">{totalCount}</span>
         <span> movies found</span>
       </div>
+
       <div className="container">
         {movies.map(movie => (
           <MovieCard
@@ -89,6 +103,7 @@ const mapStateToProps = state => {
     movies: state.movies,
     totalCount: state.totalCount,
     currentPage: state.currentPage,
+    isFetchedError: state.isFetchedError,
   };
 };
