--- my-app/src/moviesData/movies.js
@@ -1,115 +1,0 @@
-const movies = [
-  {
-    "id": 353486,
-    "title": "Jumanji: Welcome to the Jungle",
-    "tagline": "The game has evolved.",
-    "vote_average": 6.5,
-    "vote_count": 3330,
-    "release_date": "2017-12-09",
-    "poster_path": "https://image.tmdb.org/t/p/w500/bXrZ5iHBEjH7WMidbUDQ0U2xbmr.jpg",
-    "overview": "The tables are turned as four teenagers are sucked into Jumanji's world - pitted against rhinos, black mambas and an endless variety of jungle traps and puzzles. To survive, they'll play as characters from the game.",
-    "budget": 90000000,
-    "revenue": 928901353,
-    "genres": [
-      "Action",
-      "Adventure",
-      "Comedy",
-      "Family"
-    ],
-    "runtime": 119
-  },
-  {
-    "id": 399055,
-    "title": "The Shape of Water",
-    "tagline": "A Fairy Tale for Troubled Times",
-    "vote_average": 7.3,
-    "vote_count": 3200,
-    "release_date": "2017-12-01",
-    "poster_path": "https://image.tmdb.org/t/p/w500/k4FwHlMhuRR5BISY2Gm2QZHlH5Q.jpg",
-    "overview": "An other-worldly story, set against the backdrop of Cold War era America circa 1962, where a mute janitor working at a lab falls in love with an amphibious man being held captive there and devises a plan to help him escape.",
-    "budget": 19500000,
-    "revenue": 185545281,
-    "genres": [
-      "Drama",
-      "Fantasy",
-      "Romance"
-    ],
-    "runtime": 123
-  },
-  {
-    "id": 181808,
-    "title": "Star Wars: The Last Jedi",
-    "tagline": "The Saga Continues",
-    "vote_average": 7.1,
-    "vote_count": 4732,
-    "release_date": "2017-12-13",
-    "poster_path": "https://image.tmdb.org/t/p/w500/kOVEVeg59E0wsnXmF9nrh6OmWII.jpg",
-    "overview": "Rey develops her newly discovered abilities with the guidance of Luke Skywalker, who is unsettled by the strength of her powers. Meanwhile, the Resistance prepares to do battle with the First Order.",
-    "budget": 200000000,
-    "revenue": 1325937250,
-    "genres": [
-      "Fantasy",
-      "Adventure",
-      "Science Fiction"
-    ],
-    "runtime": 152
-  },
-  {
-    "id": 284054,
-    "title": "Black Panther",
-    "tagline": "Long live the king",
-    "vote_average": 7.3,
-    "vote_count": 3788,
-    "release_date": "2018-02-13",
-    "poster_path": "https://image.tmdb.org/t/p/w500/uxzzxijgPIY7slzFvMotPv8wjKA.jpg",
-    "overview": "King T'Challa returns home from America to the reclusive, technologically advanced African nation of Wakanda to serve as his country's new leader. However, T'Challa soon finds that he is challenged for the throne by factions within his own country as well as without.  Using powers reserved to Wakandan kings, T'Challa assumes the Black Panther mantel to join with girlfriend Nakia, the queen-mother, his princess-kid sister,  members of the Dora Milaje (the Wakandan \"special forces\"), and an American secret agent, to prevent Wakanda from being dragged into a world war.",
-    "budget": 200000000,
-    "revenue": 1245257672,
-    "genres": [
-      "Action",
-      "Adventure",
-      "Fantasy",
-      "Science Fiction"
-    ],
-    "runtime": 134
-  },
-  {
-    "id": 354912,
-    "title": "Coco",
-    "tagline": "The celebration of a lifetime",
-    "vote_average": 7.8,
-    "vote_count": 3619,
-    "release_date": "2017-10-27",
-    "poster_path": "https://image.tmdb.org/t/p/w500/eKi8dIrr8voobbaGzDpe8w0PVbC.jpg",
-    "overview": "Despite his family’s baffling generations-old ban on music, Miguel dreams of becoming an accomplished musician like his idol, Ernesto de la Cruz. Desperate to prove his talent, Miguel finds himself in the stunning and colorful Land of the Dead following a mysterious chain of events. Along the way, he meets charming trickster Hector, and together, they set off on an extraordinary journey to unlock the real story behind Miguel's family history.",
-    "budget": 175000000,
-    "revenue": 700920729,
-    "genres": [
-      "Adventure",
-      "Comedy",
-      "Family",
-      "Animation"
-    ],
-    "runtime": 105
-  },
-  {
-    "id": 333339,
-    "title": "Ready Player One",
-    "tagline": "A better reality awaits.",
-    "vote_average": 8.1,
-    "vote_count": 617,
-    "release_date": "2018-03-28",
-    "poster_path": "https://image.tmdb.org/t/p/w500/pU1ULUq8D3iRxl1fdX2lZIzdHuI.jpg",
-    "overview": "When the creator of a popular video game system dies, a virtual contest is created to compete for his fortune.",
-    "budget": 175000000,
-    "revenue": 0,
-    "genres": [
-      "Adventure",
-      "Science Fiction",
-      "Action"
-    ],
-    "runtime": 140
-  }
-]
-
-export default movies;
